backend/requirements.txt
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
requests>=2.31
psycopg2-binary>=2.9
gunicorn>=21.2
drf-spectacular>=0.27
django-jazzmin>=3.0
